neutron_sec_group: Avoid calling keystone.tenants.list() where possible
I was getting this error trying to update security groups from a
playbook which previously worked:

    msg: Error: You are not authorized to perform the requested action,
    admin_required. (HTTP 403)

It seems that a recent change in Icehouse means that calling
tenants.list() is now a privileged operation.

If the user is updating groups in their login tenancy we don't need to
call tenants.list() just to get the tenant ID. We can get it from the
session. This commit makes my playbook work again without requiring me
to be an admin.